Added countdown and try-catches
I had to revise the entire Timer Task class because I had to make room for a countdown that tells you how much time was remaining. I decided it was easier to just combine the two Timer Tasks into a single task, rather than have to operating at the same time. The code was also not syncing with the two timers.
Besides this, I added try-catches to actually get you to input REAL NUMBERS instead of wacky things like "asfdsad" when the program asks for a simply integer value. Silly user! The program isn't asking you for a word!
I worked on this instead of working on my English essay. I hope you enjoy it.